backend/dashboard/requirements.txt
asgiref==3.11.1
Django==5.2.11
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
psycopg2==2.9.11
PyJWT==2.11.0
sqlparse==0.5.5
typing_extensions==4.15.0
